Moved main implementation for file upload to core package and refactored files
