+++ .env.example
@@ -62,5 +62,6 @@ STORE_CONVERSATION_EVENTS=true
 DB_TEMPALTES_ENABLED=false
 
 MAX_FILE_UPLOAD_SIZE=10000 # Set the Max File Upload size; value in KB
-MIN_FILE_UPLOAD_SIZE=100 # Set the Min File Upload size; value in KB
+MIN_FILE_UPLOAD_SIZE=1 # Set the Min File Upload size; value in KB
 FILE_UPLOAD_THROTTLE_PER_MINUTE=30 # Set the rate of request to 30 requests per minute
+ENABLE_S3_FILE_UPLOAD=false
