--- app/Http/Requests/FileUploadRequest.php
@@ -1,35 +1,0 @@
-<?php
-
-namespace App\Http\Requests;
-
-use Illuminate\Foundation\Http\FormRequest;
-
-class FileUploadRequest extends FormRequest
-{
-    /**
-     * Determine if the user is authorized to make this request.
-     *
-     * @return bool
-     */
-    public function authorize()
-    {
-        return true;
-    }
-
-    /**
-     * Get the validation rules that apply to the request.
-     *
-     * @return array
-     */
-    public function rules()
-    {
-        return [
-            'file' => [
-                'file',
-                'mimes:jpg,jpeg,gif,png',
-                'max:' . config('opendialog.file_upload.max_file_upload_size'),
-                'min:' . config('opendialog.file_upload.min_file_upload_size'),
-            ]
-        ];
-    }
-}
