+++ config/filesystems.php
@@ -76,4 +76,8 @@ return [
 
     ],
 
+    'links' => [
+        public_path('uploads') => storage_path('app/uploads'),
+    ],
+
 ];
